Make project completion
Projects now work on native implementation. This works by seeing if the
todo is completed and then creating the appropriate subtask to
complete if it is. This of course means that one should never edit a
completed task, since this would create an extra task for the project.

Now also getting settings in the renamed firebase.init() function
(previously firebase.waitForAuth()).