Make week view by weekday
Week view is now shown in columns by weekday. There's a separate weekday
todo list reducer, which creates a list similar to contexts, but with
weekday names instead. Then, the css is changed based on the new `view`
attribute on the list. There's also a `week-list` css class to make both
`ha-list` and `ha-completed` behave the same in terms of css. Completed
view now reduces similarly based on completion weekday, and creates
wrapper divs for each weekday to be shown in the week grid. The week
grid is only shown on 1024px and wider width.

Contexts aren't currently visible, and there's no drag and drop yet.