Make contexts in week view
Changed the rendering of `ha-list`, `ha-completed` and `ha-todo` to be
string template based. This is a bit clearer and makes SSR a hell of a
lot easier.

Week view in list now creates a new type of grid to enable contexts to
stay on the same grid row. Added a tiny bit of css for this (grid column
flow), as well as some typographical css.

Created storybook-like preview pages for the three components mentioned
above.

Added launch configuration for Blisk browser.