Make design for meetings view
Steering meeting view via Firebase remote config parameter `meetings`.
Meetings getting logic in `lib/meetings.js`, and `ha-list` handles
getting the meetings in the today view.

Meetings are rendered as `ha-todo` elements because of css styling, but
this might be changed in the future. Meetings are identified by setting
`type = 'meeting'` on the todo, in which case it's rendered as a
meeting.

Also, removed some redundant code, mostly from Storybook stories. And
refactored `ha-todo` css to make it cleaner.